third_party/blink/web_tests/platform/linux-chrome/external/wpt/xhr/send-accept-language-expected.txt

This is a testharness.js-based test.
[FAIL] Send "sensible" default value, whatever that means
  assert_regexp_match: expected object "/Accept-Language:\s.+/" but got ""
[PASS] XMLHttpRequest: send() - Accept-Language
Harness: the test ran to completion.
